Add ability to control DB creation for mysql
Adds a new 'dbcreate' parameter to reviewboard::site to control whether or not
we should try to create a database for ReviewBoard. If this is set to false
then it is assumed that the target database has already been created by some
process outside this module. This parameter only affects mysql at this time.